refactor(activities): hoist shared wake/dim/gamepad wiring into GamepadActivityHost
Three activities (MainActivity, GamepadOverlayActivity, ConnectionsActivity)
all carried byte-identical copies of the same nine concerns: the
WakeStateController.shouldKeepScreenOn → FLAG_KEEP_SCREEN_ON collector,
the streamingSlotCount → LowPowerManager.refreshStatus collector,
setupPower(), applyScreenOn(), dispatchKeyEvent's gamepad-source gate,
dispatchGenericMotionEvent's joystick check, dispatchTouchEvent's
LowPowerTouchGate + onUserInteraction, onWindowFocusChanged's
releaseAllPhysicalReports, isGamepadSource(), and the onStop
lowPowerManager.cancel(). Plus three matching lateinit fields and two
matching @Inject declarations. About 95 lines of identical glue per
activity, 285 lines duplicated total.

This commit hoists everything into a new GamepadActivityHost in the util
package. The host:
- Constructs LowPowerManager eagerly (so dispatchTouchEvent /
  applyScreenOn can read its state without lateinit risk), with
  LowPowerManager.Views derived from the activity's root view through
  OverlayLowPowerBinding.bind().
- install() launches the two repeatOnLifecycle(STARTED) collectors on
  the activity's lifecycleScope.
- cancelDimOnStop() is called from the activity's onStop.
- dispatchKeyEvent / dispatchGenericMotionEvent / dispatchTouchEvent /
  onWindowFocusChanged return a Boolean "consumed?" signal; the
  activity falls through to super only when the host returns false.

Each activity now adds just six thin forwards (~22 lines instead of
~95) and keeps everything that's actually specific to it:
- MainActivity keeps the viewmodel observer, controller adapter,
  slot-action listener, and onOpenGamepad intent flow.
- GamepadOverlayActivity keeps the GamepadTouchView listener,
  hideSystemBars(), the status pill refresh, and the
  onGamepadStateChanged routing into BT/satellite.
- ConnectionsActivity keeps the scan/pair/connect/forget flows and
  the Bluetooth permissions/discoverability dance.

Inheritance wasn't an option because MainActivity extends GameActivity
while the other two extend AppCompatActivity directly; composition is
the only clean fit. The host takes the activity as
`AppCompatActivity` (which GameActivity inherits from) so all three
work.

No behaviour changes. All existing tests (LowPowerManagerTest,
WakeStateControllerTest, etc.) still pin the underlying logic; the host
is plumbing on top of those tested pieces. A future
GamepadActivityHostTest would need to drive the lifecycle through a
LifecycleRegistry, but the dispatch surface is already covered by the
SatelliteNative / LowPowerManager / WakeStateController unit tests
that the host calls into.

Net diff: 396 lines deleted from the three activities, 48 thin
forwards added back, 225-line host added. Overall ~75 lines fewer,
plus one obvious place to wire new activities into the same set of
behaviours.

Verification:
  ktlint ✓ · detekt ✓ · Android lint ✓ ·
  282 JVM tests / 0 failures (unchanged) ·
  55 C++ tests / 0 failures (unchanged) ·
  debug APK builds ✓
