+++ app/src/main/java/com/tinkernorth/dish/ui/main/GamepadOverlayActivity.kt
@@ -6,13 +6,11 @@ package com.tinkernorth.dish.ui.main
 import android.graphics.drawable.GradientDrawable
 import android.os.Build
 import android.os.Bundle
-import android.view.InputDevice
 import android.view.KeyEvent
 import android.view.MotionEvent
 import android.view.View
 import android.view.WindowInsets
 import android.view.WindowInsetsController
-import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
@@ -22,16 +20,13 @@ import com.tinkernorth.dish.data.network.ConnectionHub
 import com.tinkernorth.dish.data.network.ConnectionKind
 import com.tinkernorth.dish.data.network.ConnectionLive
 import com.tinkernorth.dish.data.network.SatelliteConnectionManager
-import com.tinkernorth.dish.data.network.SatelliteNative
 import com.tinkernorth.dish.data.network.WakeStateController
 import com.tinkernorth.dish.data.repository.PhysicalGamepadRegistry
 import com.tinkernorth.dish.databinding.ActivityGamepadOverlayBinding
-import com.tinkernorth.dish.databinding.OverlayLowPowerBinding
 import com.tinkernorth.dish.ui.bluetooth.BluetoothGamepadRegistry
 import com.tinkernorth.dish.ui.bluetooth.hidToXusb
 import com.tinkernorth.dish.ui.common.GamepadTouchView
-import com.tinkernorth.dish.util.LowPowerManager
-import com.tinkernorth.dish.util.LowPowerTouchGate
+import com.tinkernorth.dish.util.GamepadActivityHost
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -44,6 +39,10 @@ import javax.inject.Inject
  * or the matching [com.tinkernorth.dish.data.network.SatelliteConnection] in
  * the [SatelliteConnectionManager]. Both owners outlive the host activity so
  * the same session is reused on re-entry.
+ *
+ * Wake-lock state, dim-after-idle, and physical-gamepad pass-through all
+ * live in [GamepadActivityHost] — this activity only owns the on-screen
+ * touch gamepad and the status pill at the top of the overlay.
  */
 @AndroidEntryPoint
 class GamepadOverlayActivity :
@@ -60,18 +59,17 @@ class GamepadOverlayActivity :
     @Inject lateinit var gamepadRegistry: PhysicalGamepadRegistry
 
     private lateinit var binding: ActivityGamepadOverlayBinding
-    private lateinit var lowPowerBinding: OverlayLowPowerBinding
-    private lateinit var lowPowerManager: LowPowerManager
-    private val lowPowerTouchGate = LowPowerTouchGate()
+    private lateinit var gamepadHost: GamepadActivityHost
     private var connectionId: String = ""
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityGamepadOverlayBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        lowPowerBinding = OverlayLowPowerBinding.bind(binding.root)
+        gamepadHost =
+            GamepadActivityHost(this, binding.root, wakeState, gamepadRegistry)
+                .also { it.install() }
         hideSystemBars()
-        setupPower()
 
         connectionId = intent.getStringExtra(EXTRA_CONNECTION_ID).orEmpty()
         binding.gamepadTouchView.listener = this
@@ -88,50 +86,11 @@ class GamepadOverlayActivity :
                 hub.connections.collect { refreshStatus() }
             }
         }
-        // FLAG_KEEP_SCREEN_ON is per-window, so each activity has to flip it
-        // itself; the actual decision lives in the process-scoped
-        // WakeStateController and is shared with MainActivity.
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                wakeState.shouldKeepScreenOn.collect(::applyScreenOn)
-            }
-        }
-        // Keep the dim-overlay's count line reactive to bind/unbind — see
-        // MainActivity for the same pattern.
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                wakeState.streamingSlotCount.collect { lowPowerManager.refreshStatus() }
-            }
-        }
     }
 
     override fun onStop() {
         super.onStop()
-        // Kill the dim-screen state machine cleanly; the next onStart re-derives
-        // it from the wake-state collector.
-        lowPowerManager.cancel()
-    }
-
-    private fun setupPower() {
-        lowPowerManager = LowPowerManager(window)
-        lowPowerManager.views =
-            LowPowerManager.Views(
-                llCountdownBanner = lowPowerBinding.llCountdownBanner,
-                tvCountdownSeconds = lowPowerBinding.tvCountdownSeconds,
-                flLowPowerOverlay = lowPowerBinding.flLowPowerOverlay,
-                tvLowPowerTime = lowPowerBinding.tvLowPowerTime,
-                tvLowPowerStatus = lowPowerBinding.tvLowPowerStatus,
-            )
-        lowPowerManager.activeControllerCount = { wakeState.streamingSlotCount.value }
-    }
-
-    private fun applyScreenOn(active: Boolean) {
-        if (active) {
-            window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        } else {
-            window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        }
-        lowPowerManager.onLockStateChanged(active)
+        gamepadHost.cancelDimOnStop()
     }
 
     private fun refreshStatus() {
@@ -204,88 +163,19 @@ class GamepadOverlayActivity :
     }
 
     // ═══════════════════════════════════════════════════════════════════════
-    //  PHYSICAL INPUT DISPATCH
-    //
-    //  Mirrors MainActivity: while the overlay is foreground we still need to
-    //  feed physical gamepad events into the native pipeline so they reach
-    //  the bound satellite/BT slot instead of being absorbed by Android's
-    //  default focus-navigation handling (A = click, BACK = finish, dpad =
-    //  view traversal). The per-device bindings consulted by
-    //  SatelliteNative.process* live in the process-scoped
-    //  PhysicalSlotBindingObserver, so they stay valid across the
-    //  MainActivity → overlay handoff.
+    //  INPUT DISPATCH — forwarded to GamepadActivityHost
     // ═══════════════════════════════════════════════════════════════════════
 
-    override fun dispatchKeyEvent(event: KeyEvent): Boolean {
-        // Same shape as MainActivity: trust the device, not the event source.
-        // Generic HID joystick adapters fire button events with
-        // src=SOURCE_KEYBOARD even though the device has SOURCE_JOYSTICK; we
-        // consume every key event from a known gamepad device so Android's
-        // fallback-action pipeline can't synthesise KEYCODE_DPAD_CENTER and
-        // click views out from under the overlay.
-        val isKnownGamepad = event.deviceId in gamepadRegistry.devices.value
-        if (isGamepadSource(event.source) || isKnownGamepad) {
-            SatelliteNative.processGamepadKeyEvent(
-                event.deviceId,
-                event.source,
-                event.action,
-                event.keyCode,
-            )
-            return true
-        }
-        return super.dispatchKeyEvent(event)
-    }
+    override fun dispatchKeyEvent(event: KeyEvent): Boolean = gamepadHost.dispatchKeyEvent(event) || super.dispatchKeyEvent(event)
 
-    override fun dispatchGenericMotionEvent(event: MotionEvent): Boolean {
-        val isJoy =
-            (event.source and InputDevice.SOURCE_JOYSTICK) == InputDevice.SOURCE_JOYSTICK ||
-                event.deviceId in gamepadRegistry.devices.value
-        if (isJoy &&
-            SatelliteNative.processGamepadMotionEvent(
-                event.deviceId,
-                event.source,
-                event.action,
-                event.getAxisValue(MotionEvent.AXIS_X),
-                event.getAxisValue(MotionEvent.AXIS_Y),
-                event.getAxisValue(MotionEvent.AXIS_Z),
-                event.getAxisValue(MotionEvent.AXIS_RZ),
-                event.getAxisValue(MotionEvent.AXIS_RX),
-                event.getAxisValue(MotionEvent.AXIS_RY),
-                event.getAxisValue(MotionEvent.AXIS_HAT_X),
-                event.getAxisValue(MotionEvent.AXIS_HAT_Y),
-                event.getAxisValue(MotionEvent.AXIS_LTRIGGER),
-                event.getAxisValue(MotionEvent.AXIS_RTRIGGER),
-                event.getAxisValue(MotionEvent.AXIS_BRAKE),
-                event.getAxisValue(MotionEvent.AXIS_GAS),
-            )
-        ) {
-            return true
-        }
-        return super.dispatchGenericMotionEvent(event)
-    }
+    override fun dispatchGenericMotionEvent(event: MotionEvent): Boolean =
+        gamepadHost.dispatchGenericMotionEvent(event) || super.dispatchGenericMotionEvent(event)
 
-    private fun isGamepadSource(source: Int): Boolean =
-        (source and InputDevice.SOURCE_GAMEPAD) == InputDevice.SOURCE_GAMEPAD ||
-            (source and InputDevice.SOURCE_JOYSTICK) == InputDevice.SOURCE_JOYSTICK
-
-    override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
-        // Same shape as MainActivity: snapshot overlay-active state first so a
-        // DOWN that dismisses the dim wins the gate and the rest of the gesture
-        // is swallowed before the gamepad view sees it as a button press.
-        val overlayActive = lowPowerManager.state == LowPowerManager.State.ACTIVE
-        val consume = lowPowerTouchGate.onDispatch(ev.action, overlayActive)
-        if (ev.action == MotionEvent.ACTION_DOWN && wakeState.shouldKeepScreenOn.value) {
-            lowPowerManager.onUserInteraction()
-        }
-        return if (consume) true else super.dispatchTouchEvent(ev)
-    }
+    override fun dispatchTouchEvent(ev: MotionEvent): Boolean = gamepadHost.dispatchTouchEvent(ev) || super.dispatchTouchEvent(ev)
 
     override fun onWindowFocusChanged(hasFocus: Boolean) {
         super.onWindowFocusChanged(hasFocus)
-        // Symmetric with MainActivity: zero every bound device on focus loss
-        // so no button stays held server-side if the overlay is interrupted
-        // (notification shade, incoming call, etc.).
-        if (!hasFocus) SatelliteNative.releaseAllPhysicalReports()
+        gamepadHost.onWindowFocusChanged(hasFocus)
     }
 
     companion object {
